Added initial support for hosting Caustic::Renderer inside a WPF app. This change requires D3D11Image. Fixed some warnings in various compute shaders Fixed problem where RendererMarshaller was returning the underlying IRenderer instead of itself (i.e. it would return an IRenderer that would not perform marshalling) Added SetFinalRenderTarget to IRenderer. This allows for setting the final render target used by the renderer (needed for D3D and WPF interop) Also added ability to create offscreen window if no window is available. This allows for creating a renderer that isn't attached to a specific HWND.

+++ Libraries/Rendering/SceneGraph/MCAllocVerts.cs
@@ -31,7 +31,7 @@ globallycoherent RWStructuredBuffer<Counts> counts : register(u4);
 [numthreads(8, 8, 8)]
 void CS(uint3 DTid : SV_DispatchThreadID)
 {
-    if (DTid.x >= subdivisions || DTid.y >= subdivisions || DTid.z >= subdivisions)
+    if (DTid.x >= (uint)subdivisions || DTid.y >= (uint)subdivisions || DTid.z >= (uint)subdivisions)
         return;
     uint numCellsXY = uint(numCellsX * numCellsY);
     uint addr = DTid.x + DTid.y * numCellsX + DTid.z * numCellsXY;
