+++ Libraries/Rendering/SceneGraph/MCCountVerts.cs
@@ -101,7 +101,7 @@ uint VertexIDToVoxelAddr(uint addr, int vertexId)
 [numthreads(8, 8, 8)]
 void CS(uint3 DTid : SV_DispatchThreadID)
 {
-    if (DTid.x >= subdivisions || DTid.y >= subdivisions || DTid.z >= subdivisions)
+    if (DTid.x >= (uint)subdivisions || DTid.y >= (uint)subdivisions || DTid.z >= (uint)subdivisions)
         return;
     uint numCellsXY = uint(numCellsX * numCellsY);
     uint addr = DTid.x + DTid.y * numCellsX + DTid.z * numCellsXY;
