+++ Libraries/Rendering/SceneGraph/MCEmitVerts.cs
@@ -179,7 +179,7 @@ uint VertexIDToVoxelAddr(uint addr, int vertexId)
 [numthreads(8, 8, 8)]
 void CS(uint3 DTid : SV_DispatchThreadID)
 {
-    if (DTid.x >= subdivisions || DTid.y >= subdivisions || DTid.z >= subdivisions)
+    if (DTid.x >= (uint)subdivisions || DTid.y >= (uint)subdivisions || DTid.z >= (uint)subdivisions)
         return;
     // Determine the address of this voxel
     uint numCellsXY = numCellsX * numCellsY;
